x64: brgemm: rescope PR to brgemm_desc_t and its readers
Per review, narrow this PR to a single infrastructure structure so it is
independently buildable and testable, isolating changes to the structure
and its reader locations.

- Keep brgemm_desc_t / brgemm_attr_t field widening to dim_t.
- Revert all other structures and files that were touched only because
  they share headers (conv/binary/primitive conf structs, GPU, nvidia,
  tests). Those primitives will address their portion when their turn
  comes.
- Add minimal jit_generator add/sub/cmp dim_t overloads (enable_if gated
  on dim_t) so offset-taking Xbyak calls from widened readers compile;
  int/uint32_t call sites keep the native overloads.
- Reader adjustments limited to brgemm-directory consumers plus the
  brgemm_desc_t readers in the brgemm conv/matmul/post-ops kernels, with
  explicit boundary casts where the target field stays int.

size_t -> int cases are C4267, not the C4244 this effort targets, and are
left unchanged. Register indices and AMX tile rows/colsb stay int and are
checked against their semantic ranges rather than data-type maxima.

Verified: all cpp files in this PR compile clean under
-Wshorten-64-to-32 with no new warnings introduced vs. main; full library
builds green with WERROR=ON.

+++ src/cpu/x64/brgemm/brgemm_utils.cpp
@@ -283,11 +283,9 @@ int calculate_max_bcast_block(brgemm_desc_t *brg, const int adj_ld_block2) {
     // ----- post-ops and store accumulators -----
     const int beta_regs = !one_of(brg->beta, 1.f, 0.f);
 
-    // aux_vec_count() returns size_t but the post-ops register count is
-    // always small and bounded by the number of vector registers.
     const int postops_regs = brg->attr()
-            ? static_cast<int>(injector::aux_vec_count(
-                      brg->attr()->post_ops_, brg->isa_impl, true))
+            ? injector::aux_vec_count(
+                      brg->attr()->post_ops_, brg->isa_impl, true)
             : 0;
 
     // Emulators: fp8 emulation are supported for amx only
@@ -1059,11 +1057,9 @@ status_t brdgmm_blocking(brgemm_desc_t *brg) {
     const int compute_vregs
             = jit_brdgmm_kernel_base_t<Xbyak::Zmm>::get_compute_vmm_count(*brg);
     const int bf16_emu_vregs = brg->is_bf16_emu * 4;
-    // aux_vec_count() returns size_t but the post-ops register count is
-    // always small and bounded by the number of vector registers.
     const int postops_regs = brg->attr()
-            ? static_cast<int>(injector::aux_vec_count(
-                      brg->attr()->post_ops_, brg->isa_impl, true))
+            ? injector::aux_vec_count(
+                      brg->attr()->post_ops_, brg->isa_impl, true)
             : 0;
 
     const int max_acc_vmms = max_vregs
